Separated driver functions in drivers/
