+++ drivers/drivers/__init__.py
@@ -1,0 +1,10 @@
+""" Small functions/classes providing access to driver PES to be called from driver.py """
+
+from .harmonic import harm_driver
+
+__all__ = ["__drivers__", "harm_driver" ] 
+
+# dictionary linking strings 
+__drivers__ = {
+    "harmonic" : harm_driver
+    }
